Intel: 8.2.0: Backport LAM Qemu support for SRF
This PR includes patches which support LAM for Qemu.
This patchset is based on v8.2, while backport the patches there is not
any conflicts.
The total patch number is 2.

Linear-address masking (LAM) [1], modifies the checking that is applied to
64-bit linear addresses, allowing software to use of the untranslated
address bits for metadata and masks the metadata bits before using them as
linear addresses to access memory.

When the feature is virtualized and exposed to guest, it can be used for
efficient
address sanitizers (ASAN) implementation and for optimizations in JITs and
virtual machines.

The KVM patch series can be found in [2].

[1] Intel ISE https://cdrdv2.intel.com/v1/dl/getContent/671368
Chapter Linear Address Masking (LAM)
[2] https://lore.kernel.org/kvm/20230913124227.12574-1-binbin.wu@linux.intel.com

Upstream Commits
- https://github.com/qemu/qemu/commit/0117067131f9
- https://github.com/qemu/qemu/commit/ba6780905943

Signed-off-by: Zhiquan Li <zhiquan1.li@intel.com>